Api/requirements.txt
Django==2.2.17
djangorestframework==3.12.2
django-cors-middleware==1.5.0